App: using fmtlib causes a linking failure together with boost's graphviz classes To fix this failure a workaround is move the graphviz handling from Document.cpp to a separate source file See also: https://github.com/FreeCAD/FreeCAD/pull/8350
